requirements.txt
#pip install -r requirements.txt
Django==5.2.1
djangorestframework==3.16.0
pillow==11.2.1